requirements.txt
alabaster==0.7.12
appdirs==1.4.4
attrs==19.3.0
autoflake==1.3.1
Babel==2.8.0
bandit==1.6.2
beautifulsoup4==4.9.1
bleach==3.1.5
boto==2.49.0
boto3==1.13.22
botocore==1.16.22
bs4==0.0.1
bumpversion==0.5.3
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
coverage==5.0.3
coveralls==1.10.0
cycler==0.10.0
decorator==4.4.2
distlib==0.3.0
doc8==0.8.0
docopt==0.6.2
docutils==0.15.2
emoji==0.5.4
emoji-country-flag==1.2.1
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
flake8-2020==1.5.0
flake8-bandit==2.1.2
flake8-broken-line==0.1.1
flake8-builtins==1.4.2
flake8-comprehensions==3.2.2
flake8-debugger==3.2.1
flake8-fixme==1.1.1
flake8-isort==2.8.0
flake8-logging-format==0.6.0
flake8-mutable==1.2.0
flake8-polyfill==1.0.2
flake8-variables-names==0.0.3
Flask==1.1.2
future==0.18.2
gensim==3.8.3
gitdb==4.0.5
GitPython==3.1.3
google-api-core==1.19.0
google-auth==1.16.1
google-cloud==0.34.0
google-cloud-core==1.3.0
google-cloud-storage==1.28.1
google-cloud-vision==1.0.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
googletrans==2.4.0
grpcio==1.29.0
idna==2.9
imageio==2.8.0
imagesize==1.2.0
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
keyring==21.2.1
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
more-itertools==8.3.0
networkx==2.4
newsapi-python==0.2.6
nltk==3.5
numpy==1.18.5
opencv-python==4.2.0.34
packaging==20.4
pandas==1.0.4
pbr==5.4.5
Pillow==7.1.2
pkginfo==1.5.0.1
pluggy==0.13.1
protobuf==3.12.2
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pydocstyle==5.0.2
pyflakes==2.1.1
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.3.5
pytest-cov==2.8.1
pytest-runner==5.2
python-dateutil==2.8.1
pytz==2020.1
pyupgrade==1.26.2
PyWavelets==1.1.1
PyYAML==5.3.1
rake-nltk==1.0.4
readme-renderer==26.0
regex==2020.5.14
requests==2.23.0
requests-toolbelt==0.9.1
restructuredtext-lint==1.3.0
rsa==4.0
s3transfer==0.3.3
sacremoses==0.0.43
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
sentence-transformers==0.2.6.1
sentencepiece==0.1.91
six==1.15.0
smart-open==2.0.0
smmap==3.0.4
snowballstemmer==2.0.0
soupsieve==2.0.1
Sphinx==2.3.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
stevedore==1.32.0
testfixtures==6.14.1
threadpoolctl==2.1.0
tifffile==2020.6.3
tokenize-rt==4.0.0
tokenizers==0.7.0
toml==0.10.1
torch==1.5.0
tox==3.14.3
tqdm==4.46.0
transformers==2.11.0
twilio==6.41.0
twine==3.1.1
urllib3==1.25.9
virtualenv==20.0.21
wcwidth==0.2.2
webencodings==0.5.1
Werkzeug==1.0.1
xlrd==1.2.0
zipp==3.1.0
streamlit
ibm_watson